modified:   TCP-IP/client1.click 	modified:   TCP-IP/client2.click 	modified:   TCP-IP/receiverbuffer.cc 	modified:   TCP-IP/receiverbuffer.hh 	modified:   TCP-IP/senderbuffer.cc 	modified:   TCP-IP/senderbuffer.hh 	modified:   TCP-IP/sendertcp.cc 	modified:   TCP-IP/sendertcp.hh

+++ TCP-IP/receiverbuffer.cc
@@ -178,7 +178,7 @@ WritablePacket* ReceiverBuffer::ReadOutDataPacket(int pos){
     return packet;
 }
 
-uint8_t ReceiverBuffer::ReceiverBufferRemainSize(uint8_t s, uint8_t e){
+uint32_t ReceiverBuffer::ReceiverBufferRemainSize(uint32_t s, uint32_t e){
     if(e >= s){
         return RECEIVER_BUFFER_SIZE - 1 - (e - s);
     }
